Implement file parsing and add navigation enhancements
Two major updates were made: first, file parsing was integrated into file, user, and folder services in the backend, this is necessary because the snowflake ids are saved as i64 types which are too large to be completely represented in Javascript numbers and will round the number. Second, navigation enhancements were made in the frontend to improve usability when retrieving file and folder data. Additional changes include adding new components and modifying routes.