Added image service and handling of image files
Introduced a new image service that manages images in the application. This includes the generation of various sizes for stored images and the capability of retrieving an image by file type and format. Also enhanced the file service to handle images types that do not need optimization and improved deletion logic to remove associated formats when a file is deleted.