Add support for multiple file downloads and ZIP compression
This commit adds the capability to download multiple files at once. It also introduces the use of ZIP compression to package the downloaded files together before sending to the client. Minor changes have been made to the File and Folder services and components accordingly. The error handling has been improved with new errors for unprocessable entities.