Update HTTP headers and handle download progress
This commit introduces additional HTTP headers for cors configuration in the backend code and adjusted how the filename is obtained during file download in the frontend. It also adds functionality that allows tracking of download progress on the frontend, and the generation of file names on the backend based on actual time when the request is made.