Improve file upload with folder structure support
Added support for folder uploads and implemented new file and image processing functionality in the back-end. This includes changes to file and image services, adding display of image file previews in the front-end, and introducing the `itertools` library in the back-end for effective iteration over data. Furthermore, introduced a new Tokio runtime for image processing, allowing adjustment of the amount of worker threads via an environment variable. Fixed an issue where potential folder uploads through the drop zone were treated as file uploads.