Update UI components and implement context menu
This commit updates the user interface components of the application, particularly the move and download actions. Additionally, a context menu has been implemented to provide users with additional options. A key store was also added to detect if the shift key has been pressed. The permanent delete action in the server backend was modified to allow for permanent deletion of files.