Update UI and improve code readability
This commit revamps UI elements such as Sidebar, Context menu, and File Table. It introduces style and behaviour changes to these components to improve overall user interaction and experience. Moreover, multiple refactors are carried out to enhance readability and maintainability of the code, such as changing interfaces to types where appropriate and increasing usage of helper and utility functions.