Implement recursive folder deletion functionality
Added a new delete route to delete a folder and all its contents. This includes backend logic for recursively deleting child folders and files, as well as UI changes to allow users to choose this option. It is now possible to drag and drop files into the file list to quickly upload files. In addition, some minor fixes and enhancements, like better error handling, were introduced.