Add layout and class properties, update file and folder displays
The code has been updated to add 'layout' attribute to motion.tr components in both the FolderItem and FileItem files. Also, the class for the button element in the OperationsMenu file has been modified. Additionally, changes were made in the FileTable file to display the current number of folders, files, and the total size of the files.