Implement enhanced access control and loading UI
This commit introduces a new 'NoAccess' component to replace static strings which were previously being returned for access control checks. It significantly enhances the user interface by displaying a loading indicator while checking the user's login status, and providing buttons for navigation or login if the user does not have access.
A new loading UI has been added to the app, this will be displayed while the main chunk has to download and be evaluated.