Add Recent Files view
Introduced a new Recent Files view to display the most recently updated files. To support this, sorting functionality was disabled for this specific view. The backend was also updated to support the retrieval of recent files. This offers users a quick way to access the files they've most recently interacted with.