Add storage limit for new user and error handling in user access
This update introduces a new feature in which a storage limit is incorporated for every new user created in the register service. It also introduces error handling in user access in the frontend to enhance UX. Other modifications include updating various dependencies and moving file upload utilities to a separate component for cleaner code structure.