Add drag and drop move functionality
This commit introduces the feature to move files and folders by dragging and dropping. The movement of items is coded to only happen on desktop interfaces and is disabled for touch devices. The changes also include necessary display updates for item interaction during drag and move operations. The breadcrumbs also now serve as providers for moving items.