Update UI components and add user Minidenticon
The CSS transitions were modified to better fit the overall design. Additionally, the Minidenticon library was added to generate unique avatars for users based on their usernames. Furthermore, the UI for several components was updated, improving the user experience, for example by adding transitions to buttons and adding an "Empty" state when no notifications or operations are present.