Add user role update feature
The commit introduces the ability to update a user's role. It adds an UpdateUser permission to the list of permissions in the model/role.rs. Additionally, a new 'update_role' method has been added in services/user_service.rs to handle the SQL query for updating the role in the database. The route file has been updated to accept the new role from the admin user and utilize the new 'update_role' method.