Update sharing functionality to use ExtendedShareModelDTO
This update modifies the sharing functionality to work with ExtendedShareModelDTO instead of the standard ShareModelDTO. Additionally, some minor changes were made in the query building process in share_service.rs and properties of the ShareModel have been adjusted. This allows for the sharing API to provide more detailed information in each share instance.