Implement infinite file loading with pagination
Modified file loading to use infinite loading with pagination, improving application performance by reducing the quantity of data loaded at once. Also removed the FileUploadContent component from the file list and moved it to the new menu in the header. This was necessary as the upload drop zone would interfere with the pagination.
Adjusted some backend requests to support the pagination changes.