Add PWA support with new favicon and images
The commit includes the addition of PWA support to the application. It introduces a new favicon and images for use within the app. The manifest.json file is also added, which entails PWA configuration and specifies details about the application. Resources like the SVG logo file have been added and are referenced in the manifest file. The vite.config.ts file is amended to include PWA-related functionality.