Update various components to export as default and create homepage
Updated multiple components including Login, AccessWrapper, FileList, Register and FileListByType to export as default. This change simplifies import statements. Also implemented the base structure for a new homepage that involves creation of the hero, features and footer components. The router was updated to route '/' path to the new HomePage component. In addition, the background color in the index.css file was slightly adjusted.