Finishing Quick Share functionality, refactoring upload
Extended backend upload route function to accommodate quick sharing of files, these are placed in a dedicated folder and a public share is automatically created. The frontend was finished to support this functionality and is available through a new button on the homepage and in the header.
Refactored the backend upload route function into multiple files for increase readability and maintainability.