Disabling the default context menu everywhere, adding context for new items
Disabled the default browser context menu everywhere for consistency and user immersion. Added action for opening the context menu in an empty space in the file list, this shows the "New" options which are also available in the header.

Fixed a dark mode inconsistency in the "No Access"/"Loading" overlay.